requirements.txt
discord.py
discord.py-stubs
PyNaCl
git+https://github.com/natelatchaw/CommandRouter